+++ src/components/Project.js
@@ -1,0 +1,13 @@
+import '../styles/Project.scss';
+
+function Project() {
+    return(
+        <div className='project' id='project'>
+            <div className='mw-1200 my-0 mx-auto py-5 px-3'>
+                <h3 className='section-title text-center'>Project</h3>
+            </div>
+        </div>
+    );
+}
+
+export default Project;
